+++ resources/views/services/index.blade.php
@@ -37,11 +37,13 @@
                                 @endif
                                 @if(Auth::user()->can('services/destroy'))
                                     <td>
-                                        {!! Form::open(['method' => 'DELETE', 'action' => ['ServicesController@destroy', $service->id]]) !!}
-                                            <button class="btn-link" type="submit">
-                                                <i class="fa fa-trash"></i>
-                                            </button>
-                                        {!! Form::close() !!}
+                                        @if($service->canBeEliminated())
+                                            {!! Form::open(['method' => 'DELETE', 'action' => ['ServicesController@destroy', $service->id]]) !!}
+                                                <button class="btn-link" type="submit">
+                                                    <i class="fa fa-trash"></i>
+                                                </button>
+                                            {!! Form::close() !!}
+                                        @endif
                                     </td>
                                 @endif
                             </tr>
